requirements.txt
asgiref==3.2.10
cbor2==5.2.0
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.3.0
idna==2.10
Markdown==3.3.2
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
sqlparse==0.4.1
urllib3==1.25.11
